refactor: extract git operation output seam
