+++ src/core/git/operations.rs
@@ -10,6 +10,7 @@ use crate::core::release::changelog;
 
 use super::changes::*;
 use super::commits::*;
+use super::operation_output::{run_bulk_ids, GitOutput};
 use super::primitives::is_git_repo;
 use super::{execute_git, resolve_target};
 
@@ -26,18 +27,6 @@ const NOISY_UNTRACKED_DIRS: [&str; 8] = [
     "target",
     ".cache",
 ];
-
-#[derive(Debug, Clone, Serialize)]
-
-pub struct GitOutput {
-    pub component_id: String,
-    pub path: String,
-    pub action: String,
-    pub success: bool,
-    pub exit_code: i32,
-    pub stdout: String,
-    pub stderr: String,
-}
 
 #[derive(Debug, Clone, Serialize)]
 pub struct RepoSnapshot {
@@ -63,42 +52,6 @@ pub struct RepoBaselineSnapshot {
     pub baseline_ref: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub baseline_warning: Option<String>,
-}
-
-impl GitOutput {
-    fn from_output(id: String, path: String, action: &str, output: std::process::Output) -> Self {
-        Self {
-            component_id: id,
-            path,
-            action: action.to_string(),
-            success: output.status.success(),
-            exit_code: output.status.code().unwrap_or(1),
-            stdout: scrub_git_secrets(&String::from_utf8_lossy(&output.stdout)),
-            stderr: scrub_git_secrets(&String::from_utf8_lossy(&output.stderr)),
-        }
-    }
-}
-
-fn scrub_git_secrets(value: &str) -> String {
-    let mut scrubbed = String::with_capacity(value.len());
-    let mut rest = value;
-    const NEEDLE: &str = "x-access-token:";
-
-    while let Some(start) = rest.find(NEEDLE) {
-        let token_start = start + NEEDLE.len();
-        scrubbed.push_str(&rest[..token_start]);
-        if let Some(end) = rest[token_start..].find('@') {
-            scrubbed.push_str("[REDACTED]");
-            scrubbed.push('@');
-            rest = &rest[token_start + end + 1..];
-        } else {
-            scrubbed.push_str("[REDACTED]");
-            rest = "";
-        }
-    }
-
-    scrubbed.push_str(rest);
-    scrubbed
 }
 
 // === Changes Output Types ===
@@ -479,50 +432,6 @@ pub fn status_at(component_id: Option<&str>, path_override: Option<&str>) -> Res
     let output = execute_git(&path, &["status", "--porcelain=v1"])
         .map_err(|e| Error::git_command_failed(e.to_string()))?;
     Ok(GitOutput::from_output(id, path, "status", output))
-}
-
-fn run_bulk_ids<F>(ids: &[String], action: &str, op: F) -> BulkResult<GitOutput>
-where
-    F: Fn(&str) -> Result<GitOutput>,
-{
-    let mut results = Vec::new();
-    let mut succeeded = 0usize;
-    let mut failed = 0usize;
-
-    for id in ids {
-        match op(id) {
-            Ok(output) => {
-                if output.success {
-                    succeeded += 1;
-                } else {
-                    failed += 1;
-                }
-                results.push(ItemOutcome {
-                    id: id.clone(),
-                    result: Some(output),
-                    error: None,
-                });
-            }
-            Err(e) => {
-                failed += 1;
-                results.push(ItemOutcome {
-                    id: id.clone(),
-                    result: None,
-                    error: Some(e.to_string()),
-                });
-            }
-        }
-    }
-
-    BulkResult {
-        action: action.to_string(),
-        results,
-        summary: BulkSummary {
-            total: succeeded + failed,
-            succeeded,
-            failed,
-        },
-    }
 }
 
 /// Get git status for multiple components from JSON spec.
@@ -1457,450 +1366,4 @@ pub fn changes_project_filtered(
     }
 
     Ok(build_bulk_changes_output(&filtered, include_diff))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn is_workdir_clean_returns_true_for_clean_repo() {
-        use std::fs;
-        use tempfile::TempDir;
-
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let path = temp_dir.path();
-
-        // Initialize a git repo
-        Command::new("git")
-            .args(["init"])
-            .current_dir(path)
-            .output()
-            .expect("Failed to init git repo");
-
-        // Configure git user for commits
-        Command::new("git")
-            .args(["config", "user.email", "test@test.com"])
-            .current_dir(path)
-            .output()
-            .expect("Failed to configure git email");
-
-        Command::new("git")
-            .args(["config", "user.name", "Test User"])
-            .current_dir(path)
-            .output()
-            .expect("Failed to configure git name");
-
-        // Create a file and commit it
-        fs::write(path.join("test.txt"), "content").expect("Failed to write file");
-        Command::new("git")
-            .args(["add", "."])
-            .current_dir(path)
-            .output()
-            .expect("Failed to git add");
-
-        Command::new("git")
-            .args(["commit", "-m", "Initial commit"])
-            .current_dir(path)
-            .output()
-            .expect("Failed to commit");
-
-        // Now the repo should be clean
-        assert!(
-            super::super::is_workdir_clean(path),
-            "Expected clean repo to return true"
-        );
-    }
-
-    #[test]
-    fn is_workdir_clean_returns_false_for_dirty_repo() {
-        use std::fs;
-        use tempfile::TempDir;
-
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let path = temp_dir.path();
-
-        // Initialize a git repo
-        Command::new("git")
-            .args(["init"])
-            .current_dir(path)
-            .output()
-            .expect("Failed to init git repo");
-
-        // Create an untracked file
-        fs::write(path.join("untracked.txt"), "content").expect("Failed to write file");
-
-        // Repo should be dirty (untracked file)
-        assert!(
-            !super::super::is_workdir_clean(path),
-            "Expected dirty repo to return false"
-        );
-    }
-
-    #[test]
-    fn is_workdir_clean_returns_false_for_invalid_path() {
-        let path = std::path::Path::new("/nonexistent/path/that/does/not/exist");
-        assert!(
-            !super::super::is_workdir_clean(path),
-            "Expected invalid path to return false"
-        );
-    }
-
-    // ------------------------------------------------------------------
-    // rebase / cherry-pick / push --force-with-lease tests
-    //
-    // These shell out to real git in tempdirs. They exercise the wiring
-    // (option struct → argv → execute_git → GitOutput) end-to-end without
-    // touching the homeboy registry. `--path` keeps resolve_target out of
-    // the registry path so the tests don't need HOME isolation.
-    // ------------------------------------------------------------------
-
-    /// Create a fresh git repo with a single committed file. Returns the
-    /// TempDir (drop-cleanup) and the repo path as a String.
-    fn init_repo_with_initial_commit() -> (tempfile::TempDir, String) {
-        use std::fs;
-        let dir = tempfile::TempDir::new().expect("tempdir");
-        let path = dir.path().to_string_lossy().to_string();
-
-        Command::new("git")
-            .args(["init", "-q", "-b", "main"])
-            .current_dir(&path)
-            .output()
-            .expect("git init");
-        Command::new("git")
-            .args(["config", "user.email", "test@test.com"])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-        Command::new("git")
-            .args(["config", "user.name", "Test"])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-        fs::write(dir.path().join("README.md"), "initial\n").unwrap();
-        Command::new("git")
-            .args(["add", "."])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-        Command::new("git")
-            .args(["commit", "-q", "-m", "initial"])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-
-        (dir, path)
-    }
-
-    #[test]
-    fn changes_at_path_only_discovers_portable_component_id() {
-        use std::fs;
-        let (dir, path) = init_repo_with_initial_commit();
-        fs::write(
-            dir.path().join("homeboy.json"),
-            r#"{"id":"portable-changes"}"#,
-        )
-        .unwrap();
-
-        let out = changes_at(None, None, false, Some(&path)).expect("changes_at with --path");
-
-        assert_eq!(out.component_id, "portable-changes");
-        assert_eq!(out.path, path);
-        assert!(out.success);
-    }
-
-    #[test]
-    fn changes_at_component_and_path_trusts_both_inputs() {
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        let out = changes_at(Some("explicit-changes"), None, false, Some(&path))
-            .expect("changes_at with component and --path");
-
-        assert_eq!(out.component_id, "explicit-changes");
-        assert_eq!(out.path, path);
-        assert!(out.success);
-    }
-
-    #[test]
-    fn test_remote_tag_commit() {
-        let (_dir, path) = init_repo_with_initial_commit();
-        let remote = tempfile::TempDir::new().expect("remote tempdir");
-
-        Command::new("git")
-            .args(["init", "--bare", "-q"])
-            .current_dir(remote.path())
-            .output()
-            .expect("git init bare remote");
-        Command::new("git")
-            .args([
-                "remote",
-                "add",
-                "origin",
-                remote.path().to_str().expect("remote path"),
-            ])
-            .current_dir(&path)
-            .output()
-            .expect("git remote add");
-        Command::new("git")
-            .args(["tag", "v1.2.3"])
-            .current_dir(&path)
-            .output()
-            .expect("git tag");
-        Command::new("git")
-            .args(["push", "origin", "v1.2.3"])
-            .current_dir(&path)
-            .output()
-            .expect("git push tag");
-
-        let head = get_head_commit(&path).expect("head commit");
-        let remote_tag = remote_tag_commit(&path, "v1.2.3").expect("remote tag lookup");
-
-        assert_eq!(remote_tag.as_deref(), Some(head.as_str()));
-        assert_eq!(remote_tag_commit(&path, "v9.9.9").unwrap(), None);
-    }
-
-    #[test]
-    fn rebase_against_self_is_a_noop_success() {
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        // Rebase onto HEAD — always a no-op, exit 0.
-        let out = rebase_at(
-            None,
-            RebaseOptions {
-                onto: Some("HEAD".to_string()),
-                ..Default::default()
-            },
-            Some(&path),
-        )
-        .expect("rebase_at");
-
-        assert!(out.success, "rebase HEAD should succeed: {:?}", out.stderr);
-        assert_eq!(out.action, "rebase");
-        assert_eq!(out.path, path);
-    }
-
-    #[test]
-    fn rebase_abort_outside_of_rebase_is_an_error() {
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        // git rebase --abort with no rebase in progress fails. We surface
-        // that via GitOutput { success: false, stderr } — NOT via Err —
-        // because the caller may want to inspect the message.
-        let out = rebase_at(
-            None,
-            RebaseOptions {
-                abort: true,
-                ..Default::default()
-            },
-            Some(&path),
-        )
-        .expect("rebase_at returns Ok with failed GitOutput");
-
-        assert!(!out.success);
-        assert_ne!(out.exit_code, 0);
-        assert!(
-            out.stderr.contains("rebase") || out.stderr.contains("No rebase"),
-            "expected stderr to mention rebase: {:?}",
-            out.stderr
-        );
-    }
-
-    #[test]
-    fn cherry_pick_picks_a_commit_from_another_branch() {
-        use std::fs;
-        let (dir, path) = init_repo_with_initial_commit();
-
-        // Create a side branch with a unique commit.
-        Command::new("git")
-            .args(["checkout", "-q", "-b", "side"])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-        fs::write(dir.path().join("from-side.txt"), "side\n").unwrap();
-        Command::new("git")
-            .args(["add", "."])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-        Command::new("git")
-            .args(["commit", "-q", "-m", "side commit"])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-        let side_sha = String::from_utf8(
-            Command::new("git")
-                .args(["rev-parse", "HEAD"])
-                .current_dir(&path)
-                .output()
-                .unwrap()
-                .stdout,
-        )
-        .unwrap()
-        .trim()
-        .to_string();
-
-        // Switch back to main; pick the side commit.
-        Command::new("git")
-            .args(["checkout", "-q", "main"])
-            .current_dir(&path)
-            .output()
-            .unwrap();
-
-        let out = cherry_pick_at(
-            None,
-            CherryPickOptions {
-                refs: vec![side_sha.clone()],
-                ..Default::default()
-            },
-            Some(&path),
-        )
-        .expect("cherry_pick_at");
-
-        assert!(
-            out.success,
-            "cherry-pick should succeed: stderr={:?}",
-            out.stderr
-        );
-        assert!(
-            dir.path().join("from-side.txt").exists(),
-            "cherry-picked file should exist on main"
-        );
-    }
-
-    #[test]
-    fn cherry_pick_with_no_refs_and_no_pr_errors() {
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        // Empty refs + no PR + not continue/abort → user-facing error,
-        // NOT a `git cherry-pick` invocation.
-        let err = cherry_pick_at(None, CherryPickOptions::default(), Some(&path))
-            .expect_err("cherry_pick with empty refs should Err");
-
-        let msg = err.to_string();
-        assert!(
-            msg.contains("at least one commit ref") || msg.contains("--pr"),
-            "expected helpful error, got: {}",
-            msg
-        );
-    }
-
-    #[test]
-    fn cherry_pick_abort_outside_of_pick_is_a_failed_output() {
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        let out = cherry_pick_at(
-            None,
-            CherryPickOptions {
-                abort: true,
-                ..Default::default()
-            },
-            Some(&path),
-        )
-        .expect("cherry_pick_at returns Ok with failed GitOutput");
-
-        assert!(!out.success);
-        assert_ne!(out.exit_code, 0);
-    }
-
-    #[test]
-    fn push_options_force_with_lease_includes_flag() {
-        // We can't test the real push (no remote) but we can verify that
-        // push_at with force_with_lease=true at least invokes git with
-        // the right argv. With no remote configured, git push fails —
-        // we check stderr to confirm the flag flowed through and the
-        // failure is the expected "no remote" failure, not a wiring bug.
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        let out = push_at(
-            None,
-            PushOptions {
-                tags: false,
-                force_with_lease: true,
-                ..Default::default()
-            },
-            Some(&path),
-        )
-        .expect("push_at");
-
-        assert!(!out.success, "push without remote should fail");
-        // The failure message is from git, not us — but it should at
-        // least mention the absence of a remote / upstream rather than
-        // anything about an unknown flag.
-        assert!(
-            !out.stderr.contains("unknown option") && !out.stderr.contains("invalid argument"),
-            "--force-with-lease should be a known flag, got: {}",
-            out.stderr
-        );
-    }
-
-    #[test]
-    fn push_options_remote_url_refspec_and_strip_extraheader_push_to_bare_remote() {
-        let (_dir, path) = init_repo_with_initial_commit();
-        let remote = tempfile::TempDir::new().expect("bare remote tempdir");
-        Command::new("git")
-            .args(["init", "-q", "--bare"])
-            .current_dir(remote.path())
-            .output()
-            .expect("git init --bare");
-
-        let remote_url = remote.path().to_string_lossy().to_string();
-        let out = push_at(
-            None,
-            PushOptions {
-                remote_url: Some(remote_url.clone()),
-                refspec: Some("HEAD:refs/heads/autofix".to_string()),
-                strip_extraheader: true,
-                ..Default::default()
-            },
-            Some(&path),
-        )
-        .expect("push_at");
-
-        assert!(out.success, "push should succeed: stderr={}", out.stderr);
-        let verify = Command::new("git")
-            .args(["show-ref", "--verify", "refs/heads/autofix"])
-            .current_dir(remote.path())
-            .output()
-            .expect("git show-ref");
-        assert!(verify.status.success(), "expected autofix branch on remote");
-    }
-
-    #[test]
-    fn push_token_requires_github_remote_url() {
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        let err = push_at(
-            None,
-            PushOptions {
-                remote_url: Some("https://example.com/owner/repo".to_string()),
-                token: Some("secret-token".to_string()),
-                ..Default::default()
-            },
-            Some(&path),
-        )
-        .expect_err("non-GitHub token push should fail validation");
-
-        assert!(err.to_string().contains("https://github.com/"));
-        assert!(!err.to_string().contains("secret-token"));
-    }
-
-    #[test]
-    fn scrub_git_secrets_redacts_x_access_token_urls() {
-        let output = scrub_git_secrets(
-            "fatal: could not read https://x-access-token:ghs_secret123@github.com/owner/repo.git",
-        );
-
-        assert!(!output.contains("ghs_secret123"));
-        assert!(output.contains("https://x-access-token:[REDACTED]@github.com/owner/repo.git"));
-    }
-
-    #[test]
-    fn test_short_head_revision_at() {
-        let (_dir, path) = init_repo_with_initial_commit();
-
-        let revision = short_head_revision_at(Path::new(&path)).expect("short revision");
-
-        assert!(!revision.is_empty());
-        assert!(revision.len() <= 12, "unexpected short sha: {revision}");
-    }
 }
